// packages/dnb-eufemia/src/components/popover/types.d.ts
export type PopoverArrow = 'center' | 'top' | 'right' | 'bottom' | 'left';
export type PopoverAlign = null | 'center' | 'right' | 'left';